requirements.txt
appdirs==1.4.4
arelle-release==2.15.0
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.4.1
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
chromedriver-autoinstaller==0.6.2
colorama==0.4.6
cssselect==1.2.0
dart-fss==0.4.4
Django==4.2.5
django-ckeditor==6.7.0
django-js-asset==2.1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fake-useragent==1.2.1
h11==0.14.0
halo==0.0.31
html-table-parser==0.1.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
isodate==0.6.1
log-symbols==0.0.14
lxml==4.9.3
numpy==1.26.0
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.1.1
parse==1.19.1
Pillow==10.0.1
pycparser==2.21
pyee==8.2.2
pyparsing==3.1.1
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.8.8
requests==2.31.0
requests-html==0.10.0
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spinners==0.0.24
sqlparse==0.4.4
termcolor==2.3.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
w3lib==2.1.2
webdriver-manager==4.0.1
websockets==10.4
wsproto==1.2.0
xmltodict==0.13.0
zipp==3.17.0
